pyproject.toml
[tool.poetry]
name = "Toukabot"
version = "1.1.0"
description = ""
authors = ["chiragr39@gmail.com"]


[tool.poetry.dependencies]
flask = "^1.1.2"
"discord.py" = "^1.5.1"
"python-dotenv" = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"